Issue 137 Fix the response format in the delete and purchase endpoints

+++ service/static/js/rest_api.js
@@ -11,7 +11,6 @@ $(function () {
         $("#product_category").val(res.category);
         $("#product_description").val(res.description);
         $("#product_price").val(res.price);
-        $("#product_price_range").val("");
     }
 
     /// Clears all form fields
@@ -25,6 +24,9 @@ $(function () {
         $("#product_search_name").val("");
         $("#product_search_category").val("");
         $("#product_search_description").val("");
+        $("#product_purchase_id").val("");
+        $("#product_user_id").val("");
+        $("#product_amount").val("");
     }
 
     // Updates the flash message area
@@ -59,11 +61,13 @@ $(function () {
         });
 
         ajax.done(function(res){
+            console.log(data)
             update_form_data(res)
             flash_message("Product has been Created!")
         });
 
         ajax.fail(function(res){
+            console.log(data)
             flash_message(res.responseJSON.message)
         });
     });
@@ -88,21 +92,27 @@ $(function () {
             "price": price
         };
 
-        var ajax = $.ajax({
-                type: "PUT",
-                url: "/api/products/" + product_id,
-                contentType: "application/json",
-                data: JSON.stringify(data)
-            })
+        if (product_id) {
+            var ajax = $.ajax({
+                    type: "PUT",
+                    url: "/api/products/" + product_id,
+                    contentType: "application/json",
+                    data: JSON.stringify(data)
+                })
 
-        ajax.done(function(res){
-            update_form_data(res)
-            flash_message("Product has been Updated!")
-        });
+            ajax.done(function(res){
+                console.log(data)
+                update_form_data(res)
+                flash_message("Product has been Updated!")
+            });
 
-        ajax.fail(function(res){
-            flash_message(res.responseJSON.message)
-        });
+            ajax.fail(function(res){
+                console.log(data)
+                flash_message(res.responseJSON.message)
+            });
+        } else {
+            flash_message("Product ID can not be empty")
+        }
 
     });
 
@@ -114,45 +124,49 @@ $(function () {
 
         var product_id = $("#product_id").val();
 
-        var ajax = $.ajax({
-            type: "GET",
-            url: "/api/products/" + product_id
-        })
+        if (product_id) {
+            var ajax = $.ajax({
+                type: "GET",
+                url: "/api/products/" + product_id
+            })
 
-        ajax.done(function(res){
-            //alert(res.toSource())
-            update_form_data(res)
-            $("#search_results").empty();
-            $("#search_results").append('<table class="table-striped" cellpadding="10">');
-            var header = '<tr>'
-            header += '<th style="width:10%">ID</th>'
-            header += '<th style="width:40%">Name</th>'
-            header += '<th style="width:40%">Description</th>'
-            header += '<th style="width:40%">Category</th>'
-            header += '<th style="width:10%">Price</th></tr>'
-            $("#search_results").append(header);
-            var row = "<tr><td>"+res.id+"</td><td>"+res.name+"</td><td>"+res.description+"</td><td>"+res.category+"</td><td>"+res.price+"</td></tr>";
-            $("#search_results").append(row);
+            ajax.done(function(res){
+                //alert(res.toSource())
+                update_form_data(res)
+                $("#search_results").empty();
+                $("#search_results").append('<table class="table-striped" cellpadding="10">');
+                var header = '<tr>'
+                header += '<th style="width:10%">ID</th>'
+                header += '<th style="width:40%">Name</th>'
+                header += '<th style="width:40%">Description</th>'
+                header += '<th style="width:40%">Category</th>'
+                header += '<th style="width:10%">Price</th></tr>'
+                $("#search_results").append(header);
+                var row = "<tr><td>"+res.id+"</td><td>"+res.name+"</td><td>"+res.description+"</td><td>"+res.category+"</td><td>"+res.price+"</td></tr>";
+                $("#search_results").append(row);
 
-            $("#search_results").append('</table>');
-            flash_message("Product has been Retrieved!")
-        });
+                $("#search_results").append('</table>');
+                flash_message("Product has been Retrieved!")
+            });
 
-        ajax.fail(function(res){
-            clear_form_data()
-            $("#search_results").empty();
-            $("#search_results").append('<table class="table-striped" cellpadding="10">');
-            var header = '<tr>'
-            header += '<th style="width:10%">ID</th>'
-            header += '<th style="width:40%">Name</th>'
-            header += '<th style="width:40%">Description</th>'
-            header += '<th style="width:40%">Category</th>'
-            header += '<th style="width:10%">Price</th></tr>'
-            $("#search_results").append(header);
+            ajax.fail(function(res){
+                clear_form_data()
+                $("#search_results").empty();
+                $("#search_results").append('<table class="table-striped" cellpadding="10">');
+                var header = '<tr>'
+                header += '<th style="width:10%">ID</th>'
+                header += '<th style="width:40%">Name</th>'
+                header += '<th style="width:40%">Description</th>'
+                header += '<th style="width:40%">Category</th>'
+                header += '<th style="width:10%">Price</th></tr>'
+                $("#search_results").append(header);
 
-            $("#search_results").append('</table>');
-            flash_message(res.responseJSON.message)
-        });
+                $("#search_results").append('</table>');
+                flash_message(res.responseJSON.message)
+            });
+        } else {
+            flash_message("Product ID can not be empty")
+        }
 
     });
 
@@ -164,21 +178,25 @@ $(function () {
 
         var product_id = $("#product_id").val();
 
-        var ajax = $.ajax({
-            type: "DELETE",
-            url: "/api/products/" + product_id,
-            contentType: "application/json",
-            data: '',
-        })
+        if (product_id) {
+            var ajax = $.ajax({
+                type: "DELETE",
+                url: "/api/products/" + product_id,
+                contentType: "application/json",
+                data: '',
+            })
 
-        ajax.done(function(res){
-            clear_form_data()
-            flash_message("Product has been Deleted!")
-        });
+            ajax.done(function(res){
+                clear_form_data()
+                flash_message("Product has been Deleted!")
+            });
 
-        ajax.fail(function(res){
-            flash_message(res.responseJSON.message)
-        });
+            ajax.fail(function(res){
+                flash_message(res.responseJSON.message)
+            });
+        } else {
+            flash_message("Product ID can not be empty")
+        }
     });
 
     // ****************************************
@@ -196,21 +214,27 @@ $(function () {
             "user_id": user_id
         };
 
-        var ajax = $.ajax({
-            type: "POST",
-            url: "/api/products/" + product_id + "/purchase",
-            contentType: "application/json",
-            data: JSON.stringify(data),
-        })
+        if (product_id) {
+            var ajax = $.ajax({
+                type: "POST",
+                url: "/api/products/" + product_id + "/purchase",
+                contentType: "application/json",
+                data: JSON.stringify(data),
+            })
 
-        ajax.done(function(res){
-            clear_form_data()
-            flash_message("Product has been Purchased!")
-        });
+            ajax.done(function(res){
+                console.log(data)
+                clear_form_data()
+                flash_message("Product has been Purchased!")
+            });
 
-        ajax.fail(function(res){
-            flash_message(res.responseJSON.message)
-        });
+            ajax.fail(function(res){
+                console.log(data)
+                flash_message(res.responseJSON.message)
+            });
+        } else {
+            flash_message("Product ID can not be empty")
+        }
     });
 
     // ****************************************
